rename user_database to be the database for the whole application
