+++ mycrt-backend/src/database/getRecords.py
@@ -1,12 +1,13 @@
 import pymysql
 
-from .user_database import Base, user_repository
-from .models import *
+from .mycrt_database import db
+from .models import Capture
+from .user import User
 
-session = user_repository.db_session
+session = db.db_session
 def checkUsernameExists(username):
 	""" Function to check if a username already exists inside the database
-	
+
 		Keyword arguments:
 		username -- the username that we want to check that exists
 
@@ -22,7 +23,7 @@ def checkUsernameExists(username):
 
 def checkEmailExists(email):
 	""" Function to check if an email already exists inside the database
-	
+
 		Keyword arguments:
 		email -- the email that we want to check that exists
